+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to ScholarTrail will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2025-07-08]
+
+### Changed
+- Add tooltips to child status dots, fix edit modal headers, add Edit Child Profile modal, remove View Activity option
+
 ## [2025-07-03]
 
 ### Changed